requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
python-dotenv
httpx[http2]
pytest
pytest-asyncio
langgraph
langchain
langchain-openai
langchain-anthropic
langgraph-checkpoint-postgres
psycopg[binary,pool]
langchain-google-community
docker
tavily-python
cryptography

# ===== Data / analytics =====
googlemaps
geopandas
pyogrio
shapely
pyproj
rtree
pandas>=2.2,<3
numpy==1.26.4
pyarrow>=16,<21
fastparquet>=2024.5
folium
scikit-learn>=1.4
scipy>=1.12
matplotlib>=3.8
seaborn>=0.13
plotly>=5.22
statsmodels>=0.14
pymupdf
